Moved UI handling to utilui.h/cpp
