+++ Client/qtTeamTalk/channelstree.cpp
@@ -22,6 +22,8 @@
  */
 
 #include "channelstree.h"
+#include "utilui.h"
+
 #include <QDebug>
 #include <QStack>
 #include <QHeaderView>
@@ -65,108 +67,6 @@ extern QSettings* ttSettings;
 #define COLOR_TX_VIDEO  QColor(130,209,255)
 
 const char userMimeType[] = "application/user";
-
-bool userCanTx(int userid, StreamTypes stream_type, const Channel& chan)
-{
-    int i=0;
-    while(i<TT_TRANSMITUSERS_MAX && chan.transmitUsers[i][TT_TRANSMITUSERS_USERID_INDEX])
-    {
-        if(chan.transmitUsers[i][TT_TRANSMITUSERS_USERID_INDEX] == userid && (chan.transmitUsers[i][TT_TRANSMITUSERS_STREAMTYPE_INDEX] & stream_type))
-            return (chan.uChannelType & CHANNEL_CLASSROOM) == CHANNEL_CLASSROOM;
-        else i++;
-    }
-    return (chan.uChannelType & CHANNEL_CLASSROOM) == CHANNEL_DEFAULT;
-}
-
-bool userCanChanMessage(int userid, const Channel& chan, bool includeFreeForAll /*= false*/)
-{
-    return userCanTx(userid, STREAMTYPE_CHANNELMSG, chan) || (includeFreeForAll && userCanTx(TT_TRANSMITUSERS_FREEFORALL, STREAMTYPE_CHANNELMSG, chan));
-}
-
-bool userCanVoiceTx(int userid, const Channel& chan, bool includeFreeForAll /*= false*/)
-{
-    return userCanTx(userid, STREAMTYPE_VOICE, chan) || (includeFreeForAll && userCanTx(TT_TRANSMITUSERS_FREEFORALL, STREAMTYPE_VOICE, chan));
-}
-
-bool userCanVideoTx(int userid, const Channel& chan, bool includeFreeForAll /*= false*/)
-{
-    return userCanTx(userid, STREAMTYPE_VIDEOCAPTURE, chan) || (includeFreeForAll && userCanTx(TT_TRANSMITUSERS_FREEFORALL, STREAMTYPE_VIDEOCAPTURE, chan));
-}
-
-bool userCanDesktopTx(int userid, const Channel& chan, bool includeFreeForAll /*= false*/)
-{
-    return userCanTx(userid, STREAMTYPE_DESKTOP, chan) || (includeFreeForAll && userCanTx(TT_TRANSMITUSERS_FREEFORALL, STREAMTYPE_DESKTOP, chan));
-}
-
-bool userCanMediaFileTx(int userid, const Channel& chan, bool includeFreeForAll /*= false*/)
-{
-    return userCanTx(userid, STREAMTYPE_MEDIAFILE, chan) || (includeFreeForAll && userCanTx(TT_TRANSMITUSERS_FREEFORALL, STREAMTYPE_MEDIAFILE, chan));
-}
-
-channels_t getSubChannels(int channelid, const channels_t& channels, bool recursive /*= false*/)
-{
-    channels_t subchannels;
-    for(auto ite = channels.begin(); ite != channels.end(); ++ite)
-    {
-        if(ite.value().nParentID == channelid)
-        {
-            subchannels[ite.key()] = ite.value();
-#if QT_VERSION < QT_VERSION_CHECK(6,0,0)
-            if(recursive)
-                subchannels.unite(getSubChannels(ite.value().nChannelID, channels, recursive));
-#else
-                subchannels.insert(getSubChannels(ite.value().nChannelID, channels, recursive));
-#endif
-        }
-    }
-    return subchannels;
-}
-
-channels_t getParentChannels(int channelid, const channels_t& channels)
-{
-    channels_t parents;
-    while(channels[channelid].nParentID>0)
-    {
-        parents[channels[channelid].nParentID] = channels[channelid];
-        channelid = channels[channelid].nParentID;
-    }
-    return parents;
-}
-
-users_t getChannelUsers(int channelid, const users_t& users, const channels_t& channels, bool recursive /* = false */)
-{
-    users_t result;
-    for(auto ite=users.begin();ite!=users.end();ite++)
-    {
-        if(ite.value().nChannelID == channelid)
-            result.insert(ite.key(), ite.value());
-    }
-
-    if (recursive)
-    {
-        channels_t subs = getSubChannels(channelid, channels, true);
-        for(auto i=subs.begin();i!=subs.end();++i)
-        {
-            for (auto& u : getChannelUsers(i.key(), users, channels))
-                result.insert(u.nUserID, u);
-        }
-    }
-    return result;
-}
-
-bool isFreeForAll(StreamTypes stream_type, const int transmitUsers[][2],
-                  int max_userids = TT_TRANSMITUSERS_MAX)
-{
-    int i=0;
-    while(i<max_userids && transmitUsers[i][TT_TRANSMITUSERS_USERID_INDEX] != 0)
-    {
-        if(transmitUsers[i][TT_TRANSMITUSERS_USERID_INDEX] == TT_CLASSROOM_FREEFORALL &&
-           (transmitUsers[i][TT_TRANSMITUSERS_STREAMTYPE_INDEX] & stream_type))
-            return true;
-        i++;
-    }
-    return false;
-}
 
 ChannelsTree::ChannelsTree(QWidget* parent)
 : QTreeWidget(parent)
