+++ Client/qtTeamTalk/channelstree.h
@@ -24,26 +24,10 @@
 #ifndef CHANNELSTREE_H
 #define CHANNELSTREE_H
 
-#include <QMap>
-#include <QSet>
-#include <QVector>
-#include <QTreeWidget>
-
 #include "common.h"
 
-typedef QVector<int> userids_t;
-typedef QMap<int, Channel> channels_t;
-typedef QMap<int, User> users_t;
-
-bool userCanChanMessage(int userid, const Channel& chan, bool includeFreeForAll = false);
-bool userCanVoiceTx(int userid, const Channel& chan, bool includeFreeForAll = false);
-bool userCanVideoTx(int userid, const Channel& chan, bool includeFreeForAll = false);
-bool userCanDesktopTx(int userid, const Channel& chan, bool includeFreeForAll = false);
-bool userCanMediaFileTx(int userid, const Channel& chan, bool includeFreeForAll = false);
-
-channels_t getSubChannels(int channelid, const channels_t& channels, bool recursive = false);
-channels_t getParentChannels(int channelid, const channels_t& channels);
-users_t getChannelUsers(int channelid, const users_t& users, const channels_t& channels, bool recursive = false);
+#include <QSet>
+#include <QTreeWidget>
 
 class ChannelsTree : public QTreeWidget
 {
