+++ Client/qtTeamTalk/utilui.cpp
@@ -1,0 +1,100 @@
+/*
+ * Copyright (c) 2005-2018, BearWare.dk
+ * 
+ * Contact Information:
+ *
+ * Bjoern D. Rasmussen
+ * Kirketoften 5
+ * DK-8260 Viby J
+ * Denmark
+ * Email: contact@bearware.dk
+ * Phone: +45 20 20 54 59
+ * Web: http://www.bearware.dk
+ *
+ * This source code is part of the TeamTalk SDK owned by
+ * BearWare.dk. Use of this file, or its compiled unit, requires a
+ * TeamTalk SDK License Key issued by BearWare.dk.
+ *
+ * The TeamTalk SDK License Agreement along with its Terms and
+ * Conditions are outlined in the file License.txt included with the
+ * TeamTalk SDK distribution.
+ *
+ */
+
+#include "utilui.h"
+#include "settings.h"
+
+extern QSettings* ttSettings;
+
+void setVideoTextBox(const QRect& rect, const QColor& bgcolor,
+                     const QColor& fgcolor, const QString& text,
+                     quint32 text_pos, int w_percent, int h_percent,
+                     QPainter& painter)
+{
+    int w = w_percent / 100. * rect.width();
+    int h = h_percent / 100. * rect.height();
+
+    int x, y;
+    switch(text_pos & VIDTEXT_POSITION_MASK)
+    {
+    case VIDTEXT_POSITION_TOPLEFT :
+        x = 0; y = 0;
+        break;
+    case VIDTEXT_POSITION_TOPRIGHT :
+        x = rect.width() - w;
+        y = 0;
+        break;
+    case VIDTEXT_POSITION_BOTTOMLEFT :
+        x = 0;
+        y = rect.height() - h;
+        break;
+    case VIDTEXT_POSITION_BOTTOMRIGHT :
+    default :
+        x = rect.width() - w;
+        y = rect.height() - h;
+        break;
+    }
+
+    if(h>0 && w>0)
+    {
+        const QFont font = painter.font();
+        if(font.pixelSize() != h)
+        {
+            QFont newFont(font);
+            newFont.setPixelSize(h);
+            painter.setFont(newFont);
+        }
+        painter.fillRect(x, y, w, h, bgcolor);
+        painter.setPen(fgcolor);
+        painter.drawText(x, y, w, h, Qt::AlignHCenter | Qt::AlignCenter, text);
+
+        if(font.pixelSize() != h)
+            painter.setFont(font);
+    }
+}
+
+#if defined(Q_OS_DARWIN)
+void setMacResizeMargins(QDialog* dlg, QLayout* layout)
+{
+    QSize size = dlg->size();
+    QMargins margins = layout->contentsMargins();
+    margins.setBottom(margins.bottom()+12);
+    layout->setContentsMargins(margins);
+    size += QSize(0, 12);
+    dlg->resize(size);
+}
+#endif /* Q_OS_DARWIN */
+
+void setCurrentItemData(QComboBox* cbox, const QVariant& itemdata)
+{
+    int index = cbox->findData(itemdata);
+    if(index>=0)
+        cbox->setCurrentIndex(index);
+}
+
+QVariant getCurrentItemData(QComboBox* cbox, const QVariant& not_found/* = QVariant()*/)
+{
+    if(cbox->currentIndex()>=0)
+        return cbox->itemData(cbox->currentIndex());
+    return not_found;
+}
